fix: removed some details, cleared things with Nika
