+++ src/components/Input/index.ts
@@ -1,2 +1,2 @@
 export * from './Input';
-export * from './type';
+export * from './type.d.ts';
