+++ src/components/index.ts
@@ -1,1 +1,4 @@
 export * from './Test';
+export * from './Button';
+export * from './UserProfile';
+export * from './Input';
