--- src/hooks/Pages/SignUp/index.ts
@@ -1,1 +1,0 @@
-export * from './usePasswordLogic';
