--- src/hooks/Pages/SignUp/type.d.ts
@@ -1,5 +1,0 @@
-import { UseFormReturn } from "react-hook-form"
-
-export type PasswordLogicProps {
-  form: UseFormReturn<Inputs>
-}
