+++ src/hooks/index.ts
@@ -1,4 +1,1 @@
-export * from './useAxios';
 export * from './useGeoLoc';
-export * from './useGeoLoc';
-export * from './Pages';
