--- src/hooks/useCountdownToLogin.tsx
@@ -1,8 +1,0 @@
-import { Navigate, useNavigate } from 'react-router-dom';
-
-export const useCountdownToLogin = (time: number, navigateTo: string) => {
-  const navigate = useNavigate();
-  setTimeout(() => {
-    navigate(navigateTo);
-  }, time);
-};
