+++ src/pages/DefaultView/DefaultView.tsx
@@ -1,0 +1,10 @@
+import { Dashboard } from '@/pages';
+
+export const DefaultView = () => {
+  return (
+    <div>
+      <div className='text'>DefaultView</div>
+      <Dashboard data={undefined} />
+    </div>
+  );
+};
