+++ src/pages/index.ts
@@ -1,6 +1,6 @@
 export * from './Dashboard';
-export * from './DefaultView';
 export * from './GuestView';
 export * from './Login';
 export * from './NotFound';
 export * from './SignUp';
+export * from './DefaultView';
