+++ src/types/LocationType.d.ts
@@ -1,0 +1,4 @@
+export type LocationDataType = {
+  latitude: number;
+  longitude: number;
+};
