+++ src/types/index.ts
@@ -1,0 +1,2 @@
+export * from './FormTypes';
+export * from './LocationType';
